ci: Add Zsh CI workflow
This commit adds a new GitHub Actions workflow named "Zsh" to perform continuous integration for Zsh files in the repository. The workflow is triggered on push events to the "main" branch and for tags following the "v*.*.*" pattern. It also runs on pull requests that modify Zsh files or files in the "functions" directory. The workflow consists of two jobs:

1. "zsh-matrix": This job sets up a matrix of Zsh files based on the files in the repository. It uses the `find` command combined with `jq` to identify and set the Zsh files for the matrix.
2. "zsh-n": This job runs the Zsh linter (`zsh -n`) and the Zsh compiler (`zcompile`) for each file in the matrix created by the previous job.

This new workflow will help ensure the quality and correctness of Zsh files in the repository, providing early feedback on potential issues. It will promote better code reliability and maintainability in Zsh scripts.

Signed-off-by: Salvydas Lukosius <ss-o@users.noreply.github.com>